Cleaned up the testing
